+++ client/components/Goals/GoalsComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Button, Icon, Item, Label, Segment } from "semantic-ui-react";
+import { Button, Icon, Item, Label, Segment, Grid } from "semantic-ui-react";
 import { connect } from "react-redux";
 import { NavLink } from "react-router-dom";
 
@@ -11,16 +11,21 @@ class ItemExampleDivided extends Component {
           <Item>
             <Item.Image src="../../../../../../pictures/matthew.png" />
 
-            <Item.Content>
+            <Item.Content verticalAlign="middle">
               <Item.Header exact as={NavLink} to="/retirement/analysis">
                 See your retirement planning analysis
               </Item.Header>
               <Item.Extra>
-                <Label>Retirement</Label>
+                <Label>Retirement Goal</Label>
               </Item.Extra>
               <NavLink to="/retirement">
-                <Button>Update your retirement plan</Button>
+                <Button positive floated="right">
+                  Update your retirement plan
+                </Button>
               </NavLink>
+              <Button negative floated="right">
+                Delete retirement plan
+              </Button>
             </Item.Content>
           </Item>
         ) : null}
@@ -29,12 +34,20 @@ class ItemExampleDivided extends Component {
           <Item>
             <Item.Image src="../../../../../../pictures/matthew.png" />
 
-            <Item.Content>
+            <Item.Content verticalAlign="middle">
               <Item.Header>Emergency Planning Goal</Item.Header>
               <Item.Extra>
                 <p>Total Expense Goal: ${this.props.emergency.totalExpense}</p>
                 <Label>Emergency Planning</Label>
               </Item.Extra>
+              <NavLink to="/saveForEmergency">
+                <Button positive floated="right">
+                  Update your emergency plan
+                </Button>
+              </NavLink>
+              <Button negative floated="right">
+                Delete emergency plan
+              </Button>
             </Item.Content>
           </Item>
         ) : null}
@@ -43,14 +56,19 @@ class ItemExampleDivided extends Component {
           <Item>
             <Item.Image src="../../../../../../pictures/matthew.png" />
 
-            <Item.Content>
+            <Item.Content verticalAlign="middle">
               <Item.Header>See your housing goal</Item.Header>
               <Item.Extra>
                 <Label>Housing Goal</Label>
               </Item.Extra>
               <NavLink to="/house">
-                <Button>Update your housing goal</Button>
+                <Button positive floated="right">
+                  Update your mortgage plan
+                </Button>
               </NavLink>
+              <Button negative floated="right">
+                Delete mortgage plan
+              </Button>
             </Item.Content>
           </Item>
         ) : null}
@@ -64,13 +82,13 @@ class GoalsComponent extends Component {
   render() {
     let { retirement, goals, emergency, houseForm } = this.props;
     return (
-      <Segment>
+      <Grid.Column>
         <ItemExampleDivided
           retirement={retirement}
           houseForm={houseForm}
           emergency={emergency}
         />
-      </Segment>
+      </Grid.Column>
     );
   }
 }
@@ -80,7 +98,7 @@ const mapStateToProps = state => {
     form: state.form,
     user: state.user,
     retirement: state.retirement,
-    emergency: state.emergencyGoalReducer
+    emergency: state.emergencyGoalReducer,
   };
 };
