+++ client/components/Goals/House.js
@@ -2,19 +2,19 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { NavLink } from "react-router-dom";
 import { connect } from "react-redux";
-import Footer from "./Footer";
+import Footer from "../Footer";
 import {
   Segment,
   Form,
   Input,
   Button,
   Icon,
-  Statistic
+  Statistic,
 } from "semantic-ui-react";
-import DesktopContainer from "./AfterLogin/AfterLoginDesktopContainer";
-import MobileContainer from "./AfterLogin/AfterLoginMobileContainer";
-import { addHouseFormdetails, addHousePlan } from "../store";
-import { numberWithCommas } from "./utils";
+import DesktopContainer from "../AfterLogin/AfterLoginDesktopContainer";
+import MobileContainer from "../AfterLogin/AfterLoginMobileContainer";
+import { addHouseFormdetails, addHousePlan } from "../../store";
+import { numberWithCommas } from "../utils";
 /*eslint-disable*/
 
 const ResponsiveContainer = ({ children }) => (
@@ -25,14 +25,14 @@ const ResponsiveContainer = ({ children }) => (
 );
 
 ResponsiveContainer.propTypes = {
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 
 class House extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      showResult: false
+      showResult: false,
     };
     this.handleChangeLocal = this.handleChangeLocal.bind(this);
   }
@@ -56,7 +56,7 @@ class House extends Component {
       <ResponsiveContainer>
         <Segment
           id="howItWorksBackground"
-          style={{ padding: "1.5em" }}
+          style={{ padding: '1.5em' }}
           vertical
         >
           <i aria-hidden="true" className="home massive icon" />
@@ -150,7 +150,7 @@ class House extends Component {
 const mapStateToProps = state => {
   return {
     userId: state.user.id,
-    houseForm: state.houseForm
+    houseForm: state.houseForm,
   };
 };
 
@@ -164,7 +164,7 @@ const mapDispatchToProps = function(dispatch, ownProps) {
     handleSubmit(event, housePlan) {
       event.preventDefault();
       dispatch(addHousePlan(housePlan));
-    }
+    },
   };
 };
