+++ client/components/Goals/Retirement.js
@@ -1,11 +1,11 @@
 import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import Footer from "./Footer";
-import Loading from "./Loading";
+import Footer from "../Footer";
+import Loading from "../Loading";
 import { Grid, Segment } from "semantic-ui-react";
-import DesktopContainer from "./AfterLogin/AfterLoginDesktopContainer";
-import MobileContainer from "./AfterLogin/AfterLoginMobileContainer";
+import DesktopContainer from "../AfterLogin/AfterLoginDesktopContainer";
+import MobileContainer from "../AfterLogin/AfterLoginMobileContainer";
 
 import Step1 from "./Step1";
 import Step2 from "./Step2";
