+++ client/components/Goals/SaveForEmergencyConfirmPage.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Grid, Label, Segment, List, Button } from "semantic-ui-react";
 import { connect } from "react-redux";
-import { postEmergencyGoal } from "../store";
+import { postEmergencyGoal } from "../../store";
 
 class SaveForEmergenyConfirmPage extends Component {
   constructor(props) {
