+++ client/components/index.js
@@ -10,23 +10,17 @@ export { default as Login } from "./BeforeLoginPages/Login";
 export { default as SignUp } from "./BeforeLoginPages/SignUp";
 export { default as Navbar } from "./navbar";
 export { default as UserHome } from "./user-home";
-
+export { default as Profile } from "./Profile";
+export { default as UserCredit } from "./UserCredit";
 export { default as UserTrend } from "./Trends/UserTrend";
 export { default as Transactions } from "./Transactions/Transactions";
-export { default as AccountTable } from "./AccountTable";
-export { default as Profile } from "./Profile";
-export { default as Budget } from "./Budget";
-export {
-  default as BudgetHorizontalBarChart,
-} from "./BudgetHorizontalBarChart";
-export { default as BudgetProgress } from "./BudgetProgress";
-export { default as BudgetModal } from "./BudgetModal";
-export { default as Retirement } from "./Retirement";
-export { default as RetirementResult } from "./RetirementResult";
-export { default as UserCredit } from "./UserCredit";
-export { default as Goals } from "./Goals";
+export { default as Budget } from "./Budget/Budget";
+export { default as BudgetModal } from "./Budget/BudgetModal";
+export { default as Retirement } from "./Goals/Retirement";
+export { default as RetirementResult } from "./Goals/RetirementResult";
+export { default as Goals } from "./Goals/Goals";
 export {
   default as SaveForEmergencyMainPage,
-} from "./SaveForEmergencyMainPage";
-export { default as House } from "./house";
+} from "./Goals/SaveForEmergencyMainPage";
+export { default as House } from "./Goals/House";
 //export { Login, Signup } from "./auth-form";
