+++ public/style.css
@@ -61,18 +61,6 @@ nav a {
   display: block;
 }
 
-/* #map_election {
-  position: relative;
-  width: 500px;
-  height: 300px;
-}
-
-#map_election2 {
-  position: relative;
-  width: 500px;
-  height: 300px;
-} */
-
 #profile {
   display: table;
   margin-left: auto;
@@ -127,8 +115,12 @@ h4 {
   font-size: 2em;
 }
 
-h4 {
+h5 {
   font-size: 1.5em;
+}
+
+p {
+  font-size: 1em;
 }
 
 /* Mobile NavBar */
